fix: fix broken trivy-action pin, satisfy newer ruff rules across the gate suite
- aquasecurity/trivy-action was pinned to v0.28.0, whose composite
  action references aquasecurity/setup-trivy@v0.2.1 — a tag that no
  longer exists upstream ("Unable to resolve action"). Bumped to
  v0.36.0, which pins setup-trivy internally by SHA.
- CI installs an unpinned ruff, which picked up 0.16.0 and started
  enforcing rules this repo hadn't hit yet: BLE001 (four already-
  documented intentional blind excepts, now explicit via noqa),
  DTZ005 (naive datetime in a report filename timestamp, now UTC),
  ASYNC230 (two small local SARIF reads right after a subprocess
  completes — noqa'd rather than adding a thread hop for a negligible
  win), RUF046 (redundant int() around an already-int round()),
  RUF059 (unused unpacked tuple elements across several gates),
  ISC004 (wrap implicit string concat in a list literal), PLW1510
  (explicit check=False on cleanup/introspection subprocess.run calls
  whose return code was never checked), PYI034 (Scope.__enter__
  should return Self), SIM103 (inline a negated boolean return).
  Plus a full ruff-format pass.

+++ src/gandalf/gates/compliance.py
@@ -83,7 +83,7 @@ class ComplianceGate:
             )
 
         try:
-            pct = max(0, min(100, int(round(float(data.get("compliance", 0))))))
+            pct = max(0, min(100, round(float(data.get("compliance", 0)))))
         except (TypeError, ValueError):
             pct = 0
         score = pct / 100.0
